Move initializer script to early load
